+++ NEWS.md
@@ -1,3 +1,17 @@
+# officer 0.6.3.9000
+
+## Changes
+
+- the dataframe returned by `docx_comments()` gains a list column `para_id` containing 
+the id(s) of the parent paragraph(s). A possible breaking change is that the `commented_text` 
+is now returned as a list column to account for comments spanning multiple runs.
+The docs gain a description of the columns of the returned dataframe.
+
+## Issues
+
+- Fix. Refactor `docx_comments()` to (properly) account for comments spanning 
+multiple paragraphs, multiple comments in the same paragraph and replies. Closes #541.
+
 # officer 0.6.3
 
 ## Features